requirements.txt
streamlit
transformers
gensim
Pillow
pandas
torch
tensorflow
